[check] fix tuple name resolution
test-all
